refactor(identity)(NP-0): fix auth error paths
