+++ microservices/identity-service/src/domain/auth/errors/InvalidAccessTokenError.ts
