+++ microservices/identity-service/src/domain/auth/errors/InvalidSessionError.ts
